+++ src/Ext/TechnoType/Body.h
@@ -101,6 +101,9 @@ public:
 
 		Valueable<int> ForceWeapon_Naval_Decloaked;
 
+		Valueable<bool> Ammo_Shared;
+		Valueable<int> Ammo_Shared_Group;
+
 		struct LaserTrailDataEntry
 		{
 			ValueableIdx<LaserTrailTypeClass> idxType;
@@ -187,6 +190,8 @@ public:
 			, Death_NoAmmo { false }
 			, Death_Countdown { 0 }
 			, ForceWeapon_Naval_Decloaked { -1 }
+			, Ammo_Shared { false }
+			, Ammo_Shared_Group { -1 }
 		{ }
 
 		virtual ~ExtData() = default;