+++ docs/Whats-New.md
@@ -275,6 +275,7 @@ New:
 - Trigger Action 506 for Firing at waypoint (by FS-21)
 - New ways for self-killing objects under certaing cases (by FS-21)
 - `ForceWeapon.Naval.Decloacked` for overriding uncloaked underwater attack behavior (by FS-21)
+- Shared Ammo for transports to passengers (by FS-21)
 
 Vanilla fixes:
 - Fixed laser drawing code to allow for thicker lasers in house color draw mode (by Kerbiter, ChrisLv_CN)